feat: Add support for treating skipped answers as `null` (M2-8650, M2-6099, M2-8169) (#83)
* fix: add stronger typing to item response values

* fix: turn type helpers into type guards

* chore: add unused imports eslint plugin

* feat: f-flagged support for skipped answers ⇒ null

This change required more type massaging and definition to be explicit
about what types various reported scores are, particularly over the
course of the `evaluateScores` method.

The `treatNullAsZero` flag passed to the `ActivityEntity` constructor is
in turn passed to the `IActivityItem` entities belonging to that
activity. In the end, the only place its value actually affects
functionality is in the item's `getScore` method, which will treat
skipped answers as `null` or `0` based on its value. From then on, all
other changes in other functions were made to ensure possible values of
`0` or `null` are handled predictably.

* test: ➕ tests incl. updating fixture & JSON import

* test: add tests for `ConditionalLogicService`